requirements.txt
python=3.12.2
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==23.2.0
black==24.3.0
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.8
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
hydra-core==1.3.2
identify==2.5.35
idna==3.6
installer==0.7.0
isort==5.13.2
jaraco.classes==3.4.0
Jinja2==3.1.3
joblib==1.4.0
keyring==24.3.1
lightning-utilities==0.11.2
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nodeenv==1.8.0
numpy==1.26.4
omegaconf==2.3.0
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
pre-commit==3.7.0
prettier==0.0.7
ptyprocess==0.7.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.8.1
requests==2.31.0
requests-toolbelt==1.0.0
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.2.0
shellingham==1.5.4
six==1.16.0
sympy==1.12
threadpoolctl==3.4.0
tomlkit==0.12.4
torch==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tqdm==4.66.2
trove-classifiers==2024.4.10
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
yarl==1.9.4
